move core to routes/core
