--- v0/core/raw-cli/dash/routes.js
@@ -1,14 +1,0 @@
-const utils = require('../../../utils')
-const runcli = require('../../core-utils').runCliCmd
-
-module.exports = function(app) {
-	const commands = ['getgovernanceinfo', 'getpoolinfo', 'mnsync status', 'masternodelist']
-
-	commands.forEach(route => {
-		let spaceless = route.replace(/ /g, '-')
-		app.get('/' + spaceless, async (req, res) => {
-			const data = await runcli(route)
-			res.json(data)
-		})
-	})
-}
