+++ v0/routes/core/raw-cli/blockchain/routes.js
@@ -1,0 +1,12 @@
+const runcli = require('../../core-utils').runCliCmd
+
+module.exports = function(app) {
+	const commands = ['getblockchaininfo', 'getblockcount', 'getdifficulty', 'getspentinfo']
+
+	commands.forEach(route => {
+		app.get('/' + route, async (req, res) => {
+			const data = await runcli(route)
+			res.json(data)
+		})
+	})
+}
