+++ v0/routes/core/raw-cli/control/routes.js
@@ -1,8 +1,7 @@
-const utils = require('../../../utils')
 const runcli = require('../../core-utils').runCliCmd
 
 module.exports = function(app) {
-	const commands = ['getmininginfo']
+	const commands = ['getinfo']
 
 	commands.forEach(route => {
 		app.get('/' + route, async (req, res) => {
