+++ v0/routes/core/raw-cli/dash/routes.js
@@ -1,11 +1,11 @@
-const utils = require('../../../utils')
 const runcli = require('../../core-utils').runCliCmd
 
 module.exports = function(app) {
-	const commands = ['getinfo']
+	const commands = ['getgovernanceinfo', 'getpoolinfo', 'mnsync status', 'masternodelist']
 
 	commands.forEach(route => {
-		app.get('/' + route, async (req, res) => {
+		let spaceless = route.replace(/ /g, '-')
+		app.get('/' + spaceless, async (req, res) => {
 			const data = await runcli(route)
 			res.json(data)
 		})
