+++ v0/routes/core/raw-cli/routes.js
@@ -1,4 +1,3 @@
-const utils = require('../../utils')
 const runcli = require('../core-utils').runCliCmd
 
 module.exports = function(app) {
